[refactoring] ConsoleService 도입
