+++ TestConsole/src/main/java/TestConsole.java
@@ -22,90 +22,6 @@ public class TestConsole {
         this.scanner = new Scanner(in);
     }
 
-    public String read(int address){
-        try {
-            ProcessBuilder pb = new ProcessBuilder("java", "-jar", "ssd.jar", "R", Integer.toString(address) );
-            pb.inheritIO();
-            Process process = null;
-            process = pb.start();
-            process.waitFor();
-            System.out.println("process가 실행됐어요");
-        } catch (IOException | InterruptedException e) {
-            return "ERROR process " + e.getMessage();
-        }
-
-        List<String> lines;
-        try {
-            lines = Files.readAllLines(Paths.get("ssd_output.txt"));
-        } catch (IOException e) {
-            return "ERROR file " + e.getMessage();
-        }
-        return lines.get(0);
-    }
-
-    public static List<String> loadBlocks(String content) throws Exception {
-        //String content = Files.readString(Paths.get(filePath)).replaceAll("\\s+", ""); // 줄바꿈 제거
-
-        List<String> blocks = new ArrayList<>();
-        for (int i = 0; i < content.length(); i += 10) {
-            if (i + 10 <= content.length()) {
-                blocks.add(content.substring(i, i + 10));
-            }
-        }
-
-        return blocks;
-    }
-
-    public boolean write(int address ,String data){
-
-        //SHELL write 3 0x00000
-        try {
-            ProcessBuilder pb = new ProcessBuilder(
-                    "java", "-jar", "ssd.jar", "W", Integer.toString(address), data
-            );
-
-
-            pb.inheritIO(); // 콘솔 출력 연결
-
-            Process process = null;
-
-            process = pb.start();
-            process.waitFor();
-
-            String rtnStr = read(address);
-
-            if(data.equals(rtnStr)) return true;
-
-            process.destroy(); // 또는 process.destroyForcibly();
-            System.out.println("Process was forcefully terminated.");
-
-        } catch (Exception e) {
-            throw new RuntimeException(e);
-        }
-
-
-        return false;
-    }
-
-
-    public void fullRead(){
-
-
-    }
-
-    public boolean fullWrite(String data){
-        boolean eachResult = false;
-        for(int i=0;i<100;i++){
-            eachResult = write(i,data);
-            if(eachResult == false) return false;
-        }
-        return true;
-    }
-
-    public boolean readCompare(int LBA, String value) {
-        return value.equals(this.read(LBA));
-    }
-
 
     public void run() {
